Move ops modules into utils.ops package
