+++ src/rook/utils/ops/__init__.py
@@ -1,0 +1,6 @@
+"""Operation plumbing package used by rook compatibility layers."""
+
+from .base import Operation
+from .normalise import ResultSet, normalise
+
+__all__ = ["Operation", "ResultSet", "normalise"]
