+++ src/rook/utils/ops/consolidate.py
@@ -11,7 +11,7 @@ from loguru import logger
 
 from rook.catalog import get_catalog
 
-from .ops_helpers import ordered_dict, wrap_sequence
+from .helpers import ordered_dict, wrap_sequence
 
 
 def to_year(time_string):
